Fix Wishbone arbiter
Right now, when multiple masters want to access the bus,
access is granted to one of them, and is not revoked until
selected master has finished all of its transactions (cyc goes low).

This state causes master starvation if access is granted to high
bandwidth master, like cpu in busy loop.

This commit makes it so access to bus is revoked when pending transaction
is finished (ack and cyc are high) or when selected master is idle.

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>