liblitedram/ddr5_training: rearrange `sdram_ddr5_cs_ca_training`
Those functions are basically the same.

Only difference is that when SDRAM_PHY_ADDRESS_DELAY_CAPABLE IS NOT
defined, a warning should be printed.
And if SKIP_NO_DELAYS IS defined, then CS/CA training should be skipped.

This commit rearranges `sdram_ddr5_cs_ca_training` function to make this
more clear.

Signed-off-by: Michal Sieron <msieron@antmicro.com>