liblitedram/ddr5: CS readability changes
Renamed `CS_on_edge_detect` to `CS_in_eye`.
Previous name could be mistaken to mean "on the eye's edge".
Now it should be clearer.

Also don't encode pattern shift and the result in the return value.
Rather return the result if we are in the eye and just take an
additional parameter, through which we can return pattern shift that
worked.

In `cs_sample_prep` change default pattern from `0xa` to `0x5` to match
one used in the spec.
This way it should work as expected, where pattern shift of 0 uses
pattern `0101` and pattern shift of 1 uses `1010`.
To reflect that I renamed variables from `offset` and `pattern shift` to
`shift_0101`. Now, it explicitly shows what effect it might have.

To match `shift_0101` renaming, I also renamed `CS_find_offset` function to
`CS_should_shift_pattern`.

Signed-off-by: Michal Sieron <msieron@antmicro.com>