liblitedram/ddr5_training: clean global variables
Move global variable definitions down.
`enumerated` is now next to the function that sets it.

`WICA` is now static as it is not used outside of this file and is
defined next to other read/write leveling code.

Also make seed and serial arrays needed for read and write training
`static const` and change data type to uint8_t so it takes less space.

Signed-off-by: Michal Sieron <msieron@antmicro.com>