liblitedram/sdram_rcd: provide functions for the simulation
In simulation we use special I2CMockMaster module, which allows to not
implement I2CSlave for the RCD.
This commit adds functions that replace normal I2C sdram_rcd_read and
sdram_rcd_write functions when running a simulation.

As reads are not supported yet, this sdram_rcd_read just returns zeros.

Signed-off-by: Michal Sieron <msieron@antmicro.com>