liblitedram/sdram_rcd: fix invalid `sdram_rcd_read` usage
Code in ddr5_helpers.c was ignoring the fact that the `sdram_rcd_read`
function writes 5 bytes to the passed `data` array. 4 bytes of DWORD and
a status byte.

This commit fixes that, by changing array size from 4 to 5 in
ddr5_helpers.c and moving the status byte to the 5th byte from the 1st one.

Signed-off-by: Michal Sieron <msieron@antmicro.com>