Improve CS/CA training with multiple DRAM modules
Different modules on the same rank can be out of phase with each other
when it comes to which clock cycle they expect 0 and 1.
DRAM A can start sampling CS on CK t, and DRAM B on CK t+1.
A expects _-_-_- pattern, while B expects -_-_-_.

To solve this, CS training must:
* check both patterns (0x5, 0xA) on each dealy,
* keep track which pattern is expected by which module,
* and choose best the delays.

Previously CK was modified during CS and CA training.
It was removed, as it played with DRAM expected patterns.

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>