Improve clock gen for Xilinx FPGAs
This commit adds option to generate fractional divisors on feedback.
It improves granularity and allows for more clock combinations to be
supported by singe MMCM. So far this feature is only added to US+ MMCM
clock generation.

Option to disable reset delay for US+ PLL was also added.
It is done to allow for direct access to reset signal.
This is necessary for multi bank IO PHYs.

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>